Set up Feat and Feats much like spells. Also updated SkillUtils - findSkill() now creates the skill if one doesn't exist - since there's no reason to search for the skill if you don't want to do something with it. This means you can safely run applyClassSkills without verifying that skills with subclasses are already there.
UI needs to be updated to allow for dynamically adding skills with subtypes. Also Feat needs a toString for the UI, plus there should be a double click option for feats just like spells that displays the HTML formatted description.

Plus, there needs to be a way to parse the feat requirements from the string version contained in a Feat.