Set up a parser for feat's feat-specific prereqs
It's pretty in-depth, and should be able to handle any level of complexity involving weird parenthesis-having requirements. It uses the formatting of the database, so commas mean that both sides are needed, unless there's an or at the beginning of the last csv. Commas are checked for first, then "|", then "or", for order of operations. Parenthesis work as expected, in theory and practice so far.

Will still need to get a setup for other prereq types, but the work done for the feat-specific ones will make getting those to allow arbitrarily complex checks much easier.