Finished parser for feat prereqs.
It's a little rough for the "general" prereqs (Prereqs not limited to feat or skill requirements), but it's set up to return true on anything it doesn't understand. I figure we tell them that if it isn't there they definitely don't meet the requirements, but if they see it and aren't sure, they should check the full details of the feat for specifics.

Now there just needs to be a system for selecting feats on levelup, for showing their details in the lists, and for forcing them.

Additionally, there needs to be a method for more thouroughly searching through feats and spells.