Set up full Item searching - Refined item details - Added descriptions to all WeaponEnums.
Items can now be searched! For now the methods are in the Pathfinder class, but due to the size of the item searching method I'll likely move all searching/choosing into a SearchUtils class - or something similar.

AdventureGearEnums still need to be made. My friends have revolted from continuing to be work lackeys - it may take a little to get that done.