Set up initial creation method.
It's a little crude right now, but there's a few parts I'm happy with, so I'm putting it up as is for now, and I'll smooth it out later. Also added an Unchained Rogue class, and crudely set up the rest of the level up process that I could think of.