Removed guesswork in searching, Minor fixes
Searching now has lots of JComboBoxes that auto-populate with the options for that field. Pulling up the window now takes a little under a second to load, but it's not bad and is, in my opinion, worth it. Additionally, all name fields when searching will provide suggestions as you type based on the names available for the items you're searching.

Values in the auto-populated fields, and values in the name field will NOT update with custom items being added in. Additionally, they reflect values available for all feats, spells, or items - not just the ones available when searching. E.G. you level up, and are searching through the new spells available to you - even though you may not have Acid Arrow in the spells available, it will still pull up as a suggestion when typing in the name field.

Minor fixes included things like actually allowing selection of a custom class, which there was already the backbone for.