Vastly updated all "database" features
Searching for, creating, and modifying spells, items, and feats are all much nicer. Users are now sent directly to the search windows for these objects, instead of a list view with a search button for two separate windows required in normal use.

Class levels for spells can now be specified on spell creation and modification. Several item specifications can now be made on creation/modification.

Searches can now not only choose to require a certain boolean be true, but can require it to be false as well. E.G. The feats searched can *not* be teamwork feats.

Everything related to searching/creation/modification is a little prettier.

Removed icons from the list views in searches - having so many at once caused lag.